+++ Minio/Credentials/AssumeRoleBaseProvider.cs
@@ -20,6 +20,8 @@ using System.Net;
 using System.Text;
 using CommunityToolkit.HighPerformance;
 using Minio.DataModel;
+using Minio.DataModel.Result;
+using Minio.Handlers;
 using Minio.Helper;
 
 namespace Minio.Credentials;
