+++ Minio/Exceptions/BucketNotFoundException.cs
@@ -15,6 +15,8 @@
  * limitations under the License.
  */
 
+using Minio.DataModel.Result;
+
 namespace Minio.Exceptions;
 
 [Serializable]
