+++ Minio/Exceptions/MissingBucketReplicationConfigurationException.cs
@@ -14,6 +14,8 @@
  * limitations under the License.
  */
 
+using Minio.DataModel.Result;
+
 namespace Minio.Exceptions;
 
 [Serializable]
