+++ Minio/Helper/RegionHelper.cs
@@ -17,9 +17,10 @@
 using System.Collections.Concurrent;
 using System.Text.RegularExpressions;
 
-namespace Minio;
+namespace Minio.Helper;
 
-public static class Regions
+public static class RegionHelper
+
 {
     private static readonly Regex endpointRegex = new(@"s3[.\-](?<name>.*?)\.amazonaws\.com$",
         RegexOptions.IgnoreCase | RegexOptions.Compiled | RegexOptions.RightToLeft,
