+++ Minio/Minio.csproj
@@ -11,6 +11,7 @@
 	</PropertyGroup>
 
 	<ItemGroup>
+		<PackageReference Include="Microsoft.Extensions.Logging.Abstractions" Version="7.0.1" />
 		<PackageReference Include="System.IO.Hashing" Version="7.0.0" />
 		<PackageReference Include="System.Reactive" Version="6.0.0" />
 		<!--
